Kaustav | #8 | Bumped up rails version to 5.2.0. Implemented update member profile API functionality. | TG-8

+++ app/controllers/contacts_controller.rb
@@ -22,13 +22,15 @@ class ContactsController < ApplicationController
         contacts.each do | contact |
             contact_source = ContactSource.find_by(:name => contact[:name])
             member = Member.find(get_member_id_from_params)
-            contact_source_member_map = ContactSourceMemberMap.find_or_create_by(:member => member,
-                                                                       :contact_source => contact_source,
-                                                                       :value => contact[:value])
-            if contact_source_member_map[:value].blank?
-                contact_source_member_map.destroy
-            elsif contact_source_member_map.valid?
+            contact_source_member_map = ContactSourceMemberMap.find_or_initialize_by(:member => member,
+                                                                       :contact_source => contact_source)
+            if (not contact_source_member_map.new_record?) and not (contact[:value].blank?)
+                contact_source_member_map.update(:value => contact[:value])
+            elsif not contact[:value].blank?
+                contact_source_member_map[:value] = contact[:value]
                 contact_source_member_map.save
+            elsif contact[:value].blank?
+                contact_source_member_map.destroy unless contact_source_member_map.new_record?
             else
                 render json: {message: 'Request Error. Please check the information.', detailed_message: contact_source_member_map.errors.messages }, :status => :bad_request
                 return
